Поправил параметры методов

+++ sanic_app/api/product.py
@@ -7,8 +7,8 @@ from tortoise import transactions
 from tortoise.expressions import F
 
 from sanic_app.models import Product, ProductPydanticOut, User, Bill, Transaction, TransactionPydanticOut, \
-    ProductPydanticIn
-from sanic_app.serializers import TransactionParams
+    ProductPydanticIn, ProductModelCreate
+from sanic_app.serializers import TransactionParams, Status
 
 product = Blueprint("product", url_prefix="/product", strict_slashes=True)
 
@@ -19,27 +19,27 @@ product = Blueprint("product", url_prefix="/product", strict_slashes=True)
 @openapi.parameter("Authorization", str, "Bearer Token")
 @openapi.response(200, ProductPydanticOut, description="Product params")
 @protected()
-async def get_products(request):
+async def get_product(request):
     products = await ProductPydanticOut.from_queryset(Product.all())
     return HTTPResponse(body=products.json())
 
 
 @product.post('/', strict_slashes=False)
 @openapi.parameter("Authorization", str, "Bearer Token")
-@openapi.response(200, ProductPydanticOut, description="Product params")
+@openapi.response(200, ProductModelCreate, description="Product params")
 @openapi.definition(body={'application/json': ProductPydanticIn})
 @protected()
 @scoped('admin')
 @validate(json=ProductPydanticIn, body_argument="product_params")
 async def create_product(request, product_params: ProductPydanticIn):
     product = await Product.create(**product_params.dict())
-    product_pydantic = await ProductPydanticOut.from_tortoise_orm(product)
-    return HTTPResponse(product_pydantic.json(), content_type="application/json")
+    product_ser = await ProductModelCreate.from_tortoise_orm(product)
+    return HTTPResponse(product_ser.json(), content_type="application/json")
 
 
 @product.put('/<product_id>', strict_slashes=False)
 @openapi.parameter("Authorization", str, "Bearer Token")
-@openapi.response(200, ProductPydanticOut, description="Product params")
+@openapi.response(200, ProductModelCreate, description="Product params")
 @openapi.definition(body={'application/json': ProductPydanticIn})
 @protected()
 @scoped('admin')
@@ -47,10 +47,10 @@ async def create_product(request, product_params: ProductPydanticIn):
 async def update_product(request, product_id: int, product_params: ProductPydanticIn):
     product = await Product.filter(id=product_id).first()
     if not product:
-        pass
+        raise exceptions.NotFound("Error product id")
     product = await product.update_from_dict(product_params.dict())
-    product_pydantic = await ProductPydanticOut.from_tortoise_orm(product)
-    return HTTPResponse(product_pydantic.json(), content_type="application/json")
+    product_ser = await ProductModelCreate.from_tortoise_orm(product)
+    return HTTPResponse(product_ser.json(), content_type="application/json")
 
 
 @product.delete('/<product_id>', strict_slashes=False)
@@ -59,8 +59,10 @@ async def update_product(request, product_id: int, product_params: ProductPydant
 @scoped('admin')
 async def delete_product(request, product_id: int):
     product = await Product.filter(id=product_id).first()
+    if not product:
+        raise exceptions.NotFound("Incorrect product id")
     await product.delete()
-    return json({"status": "ok"}, status=204)
+    return HTTPResponse(Status().json(), content_type="application/json", status=204)
 
 
 @product.post('buy/', strict_slashes=False)
@@ -71,15 +73,14 @@ async def delete_product(request, product_id: int):
 @protected()
 @inject_user()
 @validate(json=TransactionParams, body_argument="transaction_params")
-async def buy_products(request, user: User, transaction_params: TransactionParams):
-    bill = await Bill.filter(id=transaction_params.bill_id,
-                             user__id=user.get('id')).first()
+async def buy_product(request, user: User, transaction_params: TransactionParams):
+    bill = await Bill.filter(id=transaction_params.bill_id, user__id=user.id).first()
     if bill is None:
-        raise exceptions.NotFound("Incorrect score")
+        raise exceptions.NotFound("Incorrect bill id")
 
     product = await Product.filter(id=transaction_params.product_id).first()
     if product is None:
-        raise exceptions.NotFound("Incorrect product")
+        raise exceptions.NotFound("Incorrect product id")
 
     if bill.balance < product.price:
         raise exceptions.NotFound("Not balance")
@@ -87,7 +88,7 @@ async def buy_products(request, user: User, transaction_params: TransactionParam
     async with transactions.in_transaction():
         bill.balance = F("balance") - product.price
         await bill.save(update_fields=['balance'])
-        transaction = await Transaction.create(bill=bill,
-                                               product=product)
+        transaction = await Transaction.create(bill=bill, product=product)
+
     transaction = await TransactionPydanticOut.from_tortoise_orm(transaction)
     return json(transaction.dict())
